+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/DetailedQcStatus.java
@@ -1,31 +1,8 @@
 package uk.ac.bbsrc.tgac.miso.core.data;
 
 import java.io.Serializable;
-import java.util.Date;
 
-import com.eaglegenomics.simlims.core.User;
-
-public interface DetailedQcStatus extends Serializable {
-
-  Long getId();
-
-  void setId(Long detailedQcStatusId);
-
-  User getCreatedBy();
-
-  void setCreatedBy(User createdBy);
-
-  Date getCreationDate();
-
-  void setCreationDate(Date creationDate);
-
-  User getUpdatedBy();
-
-  void setUpdatedBy(User updatedBy);
-
-  Date getLastUpdated();
-
-  void setLastUpdated(Date lastUpdated);
+public interface DetailedQcStatus extends Deletable, Identifiable, Serializable, Timestamped {
 
   Boolean getStatus();