+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultDetailedQcStatusService.java
@@ -1,6 +1,8 @@
 package uk.ac.bbsrc.tgac.miso.service.impl;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 
 import org.slf4j.Logger;
@@ -13,8 +15,12 @@ import com.eaglegenomics.simlims.core.User;
 import com.google.common.collect.Sets;
 
 import uk.ac.bbsrc.tgac.miso.core.data.DetailedQcStatus;
+import uk.ac.bbsrc.tgac.miso.core.store.DeletionStore;
 import uk.ac.bbsrc.tgac.miso.persistence.DetailedQcStatusDao;
 import uk.ac.bbsrc.tgac.miso.service.DetailedQcStatusService;
+import uk.ac.bbsrc.tgac.miso.service.exception.ValidationError;
+import uk.ac.bbsrc.tgac.miso.service.exception.ValidationException;
+import uk.ac.bbsrc.tgac.miso.service.exception.ValidationResult;
 import uk.ac.bbsrc.tgac.miso.service.security.AuthorizationManager;
 
 @Transactional(rollbackFor = Exception.class)
@@ -29,37 +35,79 @@ public class DefaultDetailedQcStatusService implements DetailedQcStatusService {
   @Autowired
   private AuthorizationManager authorizationManager;
 
+  @Autowired
+  private DeletionStore deletionStore;
+
   @Override
-  public DetailedQcStatus get(Long detailedQcStatus) throws IOException {
+  public DetailedQcStatus get(long detailedQcStatus) throws IOException {
     authorizationManager.throwIfUnauthenticated();
-    return detailedQcStatusDao.getDetailedQcStatus(detailedQcStatus);
+    return detailedQcStatusDao.get(detailedQcStatus);
   }
 
   @Override
-  public Long create(DetailedQcStatus detailedQcStatus) throws IOException {
+  public long create(DetailedQcStatus detailedQcStatus) throws IOException {
     authorizationManager.throwIfNonAdmin();
+    validateChange(detailedQcStatus, null);
     User user = authorizationManager.getCurrentUser();
-    detailedQcStatus.setCreatedBy(user);
-    detailedQcStatus.setUpdatedBy(user);
-    return detailedQcStatusDao.addDetailedQcStatus(detailedQcStatus);
+    detailedQcStatus.setChangeDetails(user);
+    return detailedQcStatusDao.create(detailedQcStatus);
   }
 
   @Override
-  public void update(DetailedQcStatus detailedQcStatus) throws IOException {
+  public long update(DetailedQcStatus detailedQcStatus) throws IOException {
     authorizationManager.throwIfNonAdmin();
-    DetailedQcStatus updatedDetailedQcStatus = get(detailedQcStatus.getId());
-    updatedDetailedQcStatus.setStatus(detailedQcStatus.getStatus());
-    updatedDetailedQcStatus.setDescription(detailedQcStatus.getDescription());
-    updatedDetailedQcStatus.setNoteRequired(detailedQcStatus.getNoteRequired());
+    DetailedQcStatus managed = get(detailedQcStatus.getId());
+    validateChange(detailedQcStatus, managed);
+    applyChanges(managed, detailedQcStatus);
     User user = authorizationManager.getCurrentUser();
-    updatedDetailedQcStatus.setUpdatedBy(user);
-    detailedQcStatusDao.update(updatedDetailedQcStatus);
+    managed.setChangeDetails(user);
+    return detailedQcStatusDao.update(managed);
+  }
+
+  private void validateChange(DetailedQcStatus detailedQcStatus, DetailedQcStatus beforeChange) {
+    List<ValidationError> errors = new ArrayList<>();
+
+    if (ValidationUtils.isSetAndChanged(DetailedQcStatus::getDescription, detailedQcStatus, beforeChange)
+        && detailedQcStatusDao.getByDescription(detailedQcStatus.getDescription()) != null) {
+      errors.add(new ValidationError("description", "There is already a detailed QC status with this description"));
+    }
+
+    if (!errors.isEmpty()) {
+      throw new ValidationException(errors);
+    }
+  }
+
+  private void applyChanges(DetailedQcStatus to, DetailedQcStatus from) {
+    to.setStatus(from.getStatus());
+    to.setDescription(from.getDescription());
+    to.setNoteRequired(from.getNoteRequired());
   }
 
   @Override
   public Set<DetailedQcStatus> getAll() throws IOException {
     authorizationManager.throwIfUnauthenticated();
-    return Sets.newHashSet(detailedQcStatusDao.getDetailedQcStatus());
+    return Sets.newHashSet(detailedQcStatusDao.list());
+  }
+
+  @Override
+  public DeletionStore getDeletionStore() {
+    return deletionStore;
+  }
+
+  @Override
+  public AuthorizationManager getAuthorizationManager() {
+    return authorizationManager;
+  }
+
+  @Override
+  public ValidationResult validateDeletion(DetailedQcStatus object) throws IOException {
+    ValidationResult result = new ValidationResult();
+    long usage = detailedQcStatusDao.getUsage(object);
+    if (usage > 0L) {
+      result.addError(new ValidationError(
+          "Detailed QC status '" + object.getDescription() + "' is used by " + usage + " sample" + (usage > 1L ? "s" : "")));
+    }
+    return result;
   }
 
 }